move CurrentAuthState query to shared
