+++ client/shared/src/auth.ts
@@ -1,0 +1,37 @@
+import { gql } from '@sourcegraph/http-client'
+
+import { CurrentAuthStateResult } from './graphql-operations'
+
+export const currentAuthStateQuery = gql`
+    query CurrentAuthState {
+        currentUser {
+            __typename
+            id
+            databaseID
+            username
+            avatarURL
+            email
+            displayName
+            siteAdmin
+            tags
+            url
+            settingsURL
+            organizations {
+                nodes {
+                    id
+                    name
+                    displayName
+                    url
+                    settingsURL
+                }
+            }
+            session {
+                canSignOut
+            }
+            viewerCanAdminister
+            tags
+            tosAccepted
+        }
+    }
+`
+export type AuthenticatedUser = NonNullable<CurrentAuthStateResult['currentUser']>
