+++ client/web/src/auth.ts
@@ -1,7 +1,8 @@
 import { Observable, ReplaySubject } from 'rxjs'
 import { catchError, map, mergeMap, tap } from 'rxjs/operators'
 
-import { dataOrThrowErrors, gql } from '@sourcegraph/http-client'
+import { dataOrThrowErrors } from '@sourcegraph/http-client'
+import { AuthenticatedUser as SharedAuthenticatedUser, currentAuthStateQuery } from '@sourcegraph/shared/src/auth'
 
 import { requestGraphQL } from './backend/graphql'
 import { CurrentAuthStateResult } from './graphql-operations'
@@ -14,44 +15,13 @@ import { CurrentAuthStateResult } from './graphql-operations'
  */
 export const authenticatedUser = new ReplaySubject<AuthenticatedUser | null>(1)
 
-export type AuthenticatedUser = NonNullable<CurrentAuthStateResult['currentUser']>
+export type AuthenticatedUser = SharedAuthenticatedUser
 
 /**
  * Fetches the current user, orgs, and config state from the remote. Emits no items, completes when done.
  */
 export function refreshAuthenticatedUser(): Observable<never> {
-    return requestGraphQL<CurrentAuthStateResult>(gql`
-        query CurrentAuthState {
-            currentUser {
-                __typename
-                id
-                databaseID
-                username
-                avatarURL
-                email
-                displayName
-                siteAdmin
-                tags
-                url
-                settingsURL
-                organizations {
-                    nodes {
-                        id
-                        name
-                        displayName
-                        url
-                        settingsURL
-                    }
-                }
-                session {
-                    canSignOut
-                }
-                viewerCanAdminister
-                tags
-                tosAccepted
-            }
-        }
-    `).pipe(
+    return requestGraphQL<CurrentAuthStateResult>(currentAuthStateQuery).pipe(
         map(dataOrThrowErrors),
         tap(data => authenticatedUser.next(data.currentUser)),
         catchError(() => {
